Move errors of consensus to its own module
